Close all connections after promise has been received
This is a commit in response to the application getting too many ECONNCLOSED errors. See https://github.com/patriksimek/node-mssql/issues/135

This might not help, but is difficult to test properly. If it doesn’t help, resort to using a single static connection instead as per https://github.com/patriksimek/node-mssql/issues/135#issuecomment-104312957